Rename the rest of the files to Faction
