+++ src/app/core/factions/constellation/faction.ts
