+++ src/app/core/factions/fort/faction.ts
