+++ src/app/core/factions/humans/faction.ts
