+++ src/app/core/factions/neutrals/faction.ts
