+++ src/app/core/heroes/constellation/luminous.ts
@@ -1,4 +1,4 @@
-import { constellationFraction } from '../../factions/constellation/fraction';
+import { constellationFraction } from '../../factions/constellation/faction';
 import { WishmasterItem } from '../../items/neutral';
 import { KneelingLight, RainOfFireSpell } from '../../spells/common';
 import { SummonSagittarSpell } from '../../spells/common/summon-sagittar';
