+++ src/app/core/unit-types/constellation/units.ts
@@ -1,5 +1,5 @@
 import { AssetsImages } from '../../assets';
-import { constellationFraction } from '../../factions/constellation/fraction';
+import { constellationFraction } from '../../factions/constellation/faction';
 import { spellDescrElem } from '../../ui';
 import { EssenceLeak } from './spells/essence-leak';
