+++ src/app/core/unit-types/humans/units.ts
@@ -1,5 +1,5 @@
 import { AssetsImages } from '../../assets';
-import { humansFaction } from '../../factions/humans/fraction';
+import { humansFaction } from '../../factions/humans/faction';
 import { FirebirdHealSpell } from '../../spells/common';
 import { KnightsDamageBlock } from '../../spells/common/damage-block/knights-damage-block';
 import { ProtectionAura } from '../../spells/common/protection';
